flag to enable admin apis

+++ server/app/routes/api/v1/va/triplers.js
@@ -305,26 +305,31 @@ async function deleteTripler(req, res) {
 module.exports = Router({mergeParams: true})
 .post('/triplers', (req, res) => {
   if (!req.authenticated) return _401(res, 'Permission denied.');
+  if (!ov_config.enable_admin_apis) return _403(res, 'Permission denied.');
   if (!req.admin) return _403(res, "Permission denied.");;
   return createTripler(req, res);
 })
 .put('/triplers/:triplerId', (req, res) => {
   if (!req.authenticated) return _401(res, 'Permission denied.');
+  if (!ov_config.enable_admin_apis) return _403(res, 'Permission denied.');
   if (!req.admin) return _403(res, "Permission denied.");;
   return updateTripler(req, res);
 })
 .put('/triplers/:triplerId/confirm', (req, res) => {
   if (!req.authenticated) return _401(res, 'Permission denied.');
+  if (!ov_config.enable_admin_apis) return _403(res, 'Permission denied.');
   if (!req.admin) return _403(res, "Permission denied.");;
   return confirmTripler(req, res);
 })
 .get('/triplers', (req, res) => {
   if (!req.authenticated) return _401(res, 'Permission denied.');
+  if (!ov_config.enable_admin_apis) return _403(res, 'Permission denied.');
   if (!req.admin) return _403(res, "Permission denied.");;
   return fetchAllTriplers(req, res);
 })
 .delete('/triplers/:triplerId', (req, res) => {
   if (!req.authenticated) return _401(res, 'Permission denied.');
+  if (!ov_config.enable_admin_apis) return _403(res, 'Permission denied.');
   if (!req.admin) return _403(res, "Permission denied.");;
   return deleteTripler(req, res);
 })
